requirements.txt
Django==5.2.8
djangorestframework==3.14.0
django-filter==23.5
django-environ==0.12.0
psycopg2-binary==2.9.11



